move some codes into report.py
