+++ config/train_[roi]_small_batch_v2.yaml
@@ -32,7 +32,7 @@ soft_labels_filename: ""
 log_dir: logs_submit
 sample_num: 6
 debug: true
-clean_debug: false
+clean_debug: true
 # normalization:
 #   mean: [0.4914, 0.4822, 0.4465]
 #   std: [0.247, 0.243, 0.261]
