+++ utils/__init__.py
@@ -1,3 +1,4 @@
 import imp
 from .common import *
 from .logs import *
+from .report import *
