+++ utils/constant.py
@@ -1,0 +1,10 @@
+CLASS_NAMES = [
+    'OK', 'AirRoomShake', 'Dead', 'Empty', 'NoAirRoom', 'Split', 'Weak',
+    'Flower'
+]
+
+BN_CLASS_NAMES = ['OK', 'NoOK']
+
+HEADER_NAMES = ['filename'] + CLASS_NAMES
+
+THRESH = [0.9, 0.8, 0.7, 0.8, 0.7, 0.6, 0.5, 0.4, 0.3, 0.2, 0.1]
